requirements.txt
Pillow
tqdm
opencv-python
mediapipe
numpy